#44-rutas y autorización

+++ app/Http/Controllers/SpecialtyController.php
@@ -17,17 +17,20 @@ public function index()
     }
     public function store(Request $request)
     {
+        $this->authorize('create', Specialty::class);
         $specialties = Specialty::create($request->all());
 
         return response()->json($specialties, 201);
     }
     public function update(Request $request, Specialty $specialties)
     {
+        $this->authorize('update', $specialties);
         $specialties->update($request->all());
         return response()->json($specialties, 200);
     }
     public function delete(Specialty $specialties)
     {
+        $this->authorize('update', $specialties);
         $specialties->delete();
 
         return response()->json(null, 204);
